% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiVN.R
\docType{package}
\name{gdpm}
\alias{gdpm}
\alias{gdpm-package}
\title{epiVN : Vaccination coverages from the WHO's EPI}
\description{
Contains the vaccine coverages data from the WHO's Expanded Programme on
Immunization.
}

