% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{je}
\alias{je}
\title{Vaccine coverage against Japanese Encephalitis in Vietnam}
\format{A data frame of 2560 rows and 7 variables:
\itemize{
   \item \code{province}: name of the province
   \item \code{year}: year of vaccination
   \item \code{nb_districts}: number of districts of the province where the vaccination was impleme
   \item \code{total_number}: total number of districts in the province
   \item \code{nb_doses}: number of doses delivered par individual
   \item \code{achieved}: total number of doses delivered in the province
   \item \code{target}: targeted total number of doses in the province
}
5 9 18}
\source{
WHO's Expanded Programme of Immunization.
}
\usage{
data(diseases)
}
\description{
A dataset containing the vaccine coverage against Japanese Encephalitis from
1997 to 2016.
}
\keyword{datasets}
